feat: Restructure permits feature and update core components
- Add new permits feature directory
- Create reusable card and app bar widgets
- Update core router and dependencies
- Remove deprecated beekeeping permits screen
- Modify map screen implementation
