+++ lib/core/widgets/custom_sliver_app_bar.dart
@@ -1,0 +1,95 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
+import 'package:royal_reserve/core/core.dart';
+
+class CustomSliverAppBar extends StatelessWidget {
+  const CustomSliverAppBar({
+    super.key,
+    required this.title,
+    required this.bgImage,
+  });
+  final String title;
+  final String bgImage;
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverAppBar(
+      expandedHeight: 100.h,
+      automaticallyImplyLeading: false,
+      flexibleSpace: FlexibleSpaceBar(
+        background: Stack(
+          fit: StackFit.expand,
+          children: [
+            /// Background Image
+            Stack(
+              children: [
+                Container(
+                  decoration: BoxDecoration(
+                    borderRadius: const BorderRadius.only(
+                      bottomLeft: Radius.circular(25),
+                      bottomRight: Radius.circular(25),
+                    ),
+                    image: DecorationImage(
+                      image: AssetImage(bgImage),
+                      fit: BoxFit.fill,
+                    ),
+                  ),
+                ),
+                Container(
+                  decoration: BoxDecoration(
+                    borderRadius: const BorderRadius.only(
+                      bottomLeft: Radius.circular(25),
+                      bottomRight: Radius.circular(25),
+                    ),
+                    color: Colors.black.withOpacity(0.3),
+                  ),
+                ),
+              ],
+            ),
+
+            /// Logo and Menu Icon - Always Visible with SafeArea Padding
+            Positioned(
+              top: MediaQuery.paddingOf(context).top + 8.w,
+              left: 16.w,
+              right: 16.w,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Image.asset(
+                    AppImages.logoWithText,
+                    width: 100.w,
+                  ),
+                  InkWell(
+                    onTap: () {
+                      context.goNamed(AppRoutes.home);
+                    },
+                    child: const Icon(
+                      Icons.arrow_forward,
+                      color: Colors.white,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+
+            /// Text in Center
+            Positioned(
+              bottom: 16.0,
+              left: 16.0,
+              right: 16.0,
+              child: Text(
+                title,
+                style: AppTextStyles.heading2.copyWith(
+                  color: AppColors.white,
+                  fontSize: 16.sp,
+                ),
+                textAlign: TextAlign.center,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
