+++ lib/features/permits/data/model/bee_regulation.dart
@@ -1,0 +1,9 @@
+class BeeRegulation {
+  final String text;
+  final bool hasBackground;
+
+  const BeeRegulation({
+    required this.text,
+    required this.hasBackground,
+  });
+}
