+++ lib/features/permits/data/remote_data/bee_regulations.dart
@@ -1,0 +1,117 @@
+// List of all regulations
+import '../model/bee_regulation.dart';
+
+class BeeRegulations {
+  static const List<BeeRegulation> regulations = [
+    BeeRegulation(
+      text: 'يمنع وضع المناحل المتنقلة داخل الروضات.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text:
+          'الالتزام باستخدام الطرق الترابية الحالية وعدم شق طريق جديدة للحفاظ على الغطاء النباتي.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'الحفاظ على مستوى نظافة المنطقة وتلافي رمي المخلفات بكافة أنواعها.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'عدم التعرض للحياة الفطرية داخل المحمية.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text:
+          'عدم العبث أو قطع الأشجار أو الأعشاب والالتزام بالحفاظ على الغطاء النباتي.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'عدم القيام بأي أعمال تؤدي إلى تغيير طبيعة أرض المحمية.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'عدم إشعال النار في المحمية بدون موقد.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'الالتزام بارتداء البذور ووسائل الحماية للوقاية من لسعات النحل.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'يتم نقل المناحل بطريقة آمنة تضمن عدم انتشار النحل في المناطق.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text:
+          'يلتزم صاحب المنحل بالمواقع المحددة من قبل الهيئة وعدم تغيير الموقع إلا بعد أخذ الموافقة المسبقة من الهيئة.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'عدم استخدام أرض المحمية لأي نشاط آخر غير مرخص به.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text:
+          'يجب وضع اللوحات التحذيرية حول المناحل والالتزام بوضع نسخة من التصريح الصادر من الهيئة على المنحل.',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text:
+          'يحق للهيئة اتخاذ كافة الإجراءات النظامية في حالة عدم الالتزام بالضوابط والاشتراطات فيجوز للهيئة إلغاء التصريح واتخاذ الإجراءات النظامية اللازمة.',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text:
+          'يتحمل صاحب المنحل كامل المسؤولية تجاه أي ضرر يحدث تجاه الأشخاص أو الممتلكات ناجحة عن المنحل و/أو الأدوات الخاصة به بدون أي تبعات قانونية أو مالية على الهيئة.',
+      hasBackground: false,
+    ),
+  ];
+
+  static const List<BeeRegulation> offDutyWorkPermitRegulations = [
+    BeeRegulation(
+      text: 'منع الصيد',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'منع الاحتطاب',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'منع التخيم',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'منع الرعي',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'منع دخول الكرافانات',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'منع رمي المخلفات',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'منع التنقل برفقة الطيور الجارحة',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text: 'منع اشعال النار على الأرض مباشرة',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'منع عبور الأبل بدون اخذ تصريح',
+      hasBackground: true,
+    ),
+    BeeRegulation(
+      text:
+          'منع استخدام أي وزيوت أو مواد كيميائية أو الآلات التي تخرج الغازات السامة',
+      hasBackground: false,
+    ),
+    BeeRegulation(
+      text: 'منع فتح مسارات جديدة في مناطق الغطاء النباتي',
+      hasBackground: true,
+    ),
+  ];
+}
