+++ lib/features/permits/widgets/row_form.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class RowForm extends StatelessWidget {
+  const RowForm({
+    super.key,
+    required this.items,
+  });
+
+  final List<Widget> items;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: items,
+    );
+  }
+}
